Require integer ids in SyncComputedSppTotalsResultSchema
updatedPlayerIds are database serial ids, which are always integers;
z.number() without .int() would silently accept a fractional value.

PR review feedback for #380.

+++ packages/api-contract/src/schemas/player.ts
@@ -37,7 +37,7 @@ export const SyncComputedSppTotalsSchema = z.object({
 });
 
 export const SyncComputedSppTotalsResultSchema = z.object({
-  updatedPlayerIds: z.array(z.number()),
+  updatedPlayerIds: z.array(z.number().int()),
 });
 
 export type Player = z.infer<typeof PlayerSchema>;
